Move ListItemDropdown to public components
