+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export * from './HeaderNavigation';
 export * from './OrdinoLogo';
 export * from './OrdinoFooter';
-
+export * from './ListItemDropdown';
