+++ src/internal/components/common/index.ts
@@ -1,2 +1,1 @@
-export * from './ListItemDropdown';
 export * from './OrdinoScrollspy';
